cloud_functions/reddit_harvester/requirements.txt
functions-framework>=3.0.0
google-cloud-bigquery>=3.0.0
google-cloud-aiplatform>=1.0.0
praw>=7.7.0
pyahocorasick>=2.0.0
google-cloud-storage>=2.0.0
